+++ elk/__init__.py
@@ -1,2 +1,2 @@
-from . import pipeline
+from . import ensemble
 from ._version import __version__
